+++ tests/integration/zia/conftest.py
@@ -15,7 +15,6 @@
 # OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 
 import os
-from functools import wraps
 
 from zscaler.zia import ZIAClientHelper
 
@@ -45,21 +44,3 @@ class MockZIAClient(ZIAClientHelper):
                 cache=None,
                 fail_safe=False,
             )
-
-
-def stub_sleep(func):
-    """Decorator to speed up time.sleep function used in any methods under test."""
-    import time
-    from time import sleep
-
-    def newsleep(seconds):
-        sleep_speed_factor = 10.0
-        sleep(seconds / sleep_speed_factor)
-
-    time.sleep = newsleep
-
-    @wraps(func)
-    def wrapper(*args, **kwargs):
-        return func(*args, **kwargs)
-
-    return wrapper